refactor(tape): remove legacy compatibility re-exports
The seven `@deprecated` modules under `src/main/session/data/`
(`tape`, `tapeEffectiveView`, `tapeFacts`, `tapeViewManifest` and the
three `tables/deepchatTape*` files) only re-exported `src/main/tape/`
symbols. Production code stopped importing them after the layering
refactor; eight test files were the last consumers and now import the
canonical modules directly.

`layerBoundaries.test.ts` drops the frozen export-contract map, the
"production must not import legacy paths" guard and its negative
fixtures, all of which only exist to police the deleted files. The two
storage-boundary exceptions for the deleted table modules go with them,
otherwise the stale-exception assertion fails. The remaining guards
(domain purity, Memory route port, concrete facade isolation, physical
storage boundary) are unchanged.

`tape-system.md` and `session-management.md` now state that Tape has no
compatibility path under `session/data`. The historical
`tape-layering` SDD is left as written.
