--- src/main/session/data/tables/deepchatTapeEffectiveSemantics.ts
@@ -1,16 +1,0 @@
-/** @deprecated Import Tape semantics from `@/tape/domain/effectiveSemantics`. */
-export {
-  messageRecordHasFinalToolUse,
-  parseAssistantBlocks,
-  parseNestedTapeJsonObject,
-  parseTapeJsonObject,
-  readTapeMessageRetractionId,
-  readTapeToolIdentity,
-  readTapeToolStatus,
-  tapeEntryToMessageRecord,
-  tapeMessageRank,
-  tapeToolRank
-} from '@/tape/domain/effectiveSemantics'
-
-/** @deprecated Import Tape semantic types from `@/tape/domain/effectiveSemantics`. */
-export type { DeepChatTapeToolIdentity } from '@/tape/domain/effectiveSemantics'
