--- src/main/session/data/tables/deepchatTapeSearchProjection.ts
@@ -1,14 +1,0 @@
-/** @deprecated Import the SQLite projection from `@/tape/infrastructure/sqlite/tapeSearchProjectionStore`. */
-export {
-  DEEPCHAT_TAPE_SEARCH_PROJECTION_VERSION,
-  DeepChatTapeSearchProjectionTable
-} from '@/tape/infrastructure/sqlite/tapeSearchProjectionStore'
-
-/** @deprecated Import projection types from `@/tape/infrastructure/sqlite/tapeSearchProjectionStore`. */
-export type {
-  DeepChatTapeSearchProjectionInput,
-  DeepChatTapeSearchProjectionMeta,
-  DeepChatTapeSearchProjectionReadResult,
-  DeepChatTapeSearchProjectionResultRow,
-  DeepChatTapeSearchProjectionRow
-} from '@/tape/infrastructure/sqlite/tapeSearchProjectionStore'
