--- src/main/session/data/tape.ts
@@ -1,19 +1,0 @@
-/** @deprecated Import Tape application APIs from `@/tape/application/sessionTape`. */
-export {
-  AgentTapeViewError,
-  normalizeSubagentTapeLinkInput,
-  normalizeTapeHandoffState,
-  SessionTape
-} from '@/tape/application/sessionTape'
-
-/** @deprecated Import Tape application types from `@/tape/application/sessionTape`. */
-export type {
-  AgentTapeViewErrorCode,
-  TapeAnchorResult,
-  TapeBackfillResult,
-  TapeForkHandle,
-  TapeInfo,
-  TapeMigrationState,
-  TapeSearchResult,
-  TapeViewManifestAssemblySources as TapeViewManifestSourceMaps
-} from '@/tape/application/sessionTape'
