--- src/main/session/data/tapeEffectiveView.ts
@@ -1,9 +1,0 @@
-/** @deprecated Import effective Tape helpers from `@/tape/domain/effectiveView`. */
-export {
-  buildEffectiveTapeView,
-  getLastEffectiveTokenUsage,
-  searchEffectiveTapeRows
-} from '@/tape/domain/effectiveView'
-
-/** @deprecated Import effective Tape types from `@/tape/domain/effectiveView`. */
-export type { EffectiveMessageEntry, EffectiveTapeView } from '@/tape/domain/effectiveView'
