--- src/main/session/data/tapeFacts.ts
@@ -1,14 +1,0 @@
-/** @deprecated Import Tape fact helpers from `@/tape/application/factPersistence`. */
-export {
-  appendMessageRecordToTape,
-  appendMessageReplacementToTape,
-  appendMessageRetractionToTape,
-  appendTapeToolFact,
-  appendToolFactsToTape,
-  buildTapeToolFactInputs,
-  tapeEntriesToEffectiveMessageRecords,
-  tapeEntryToMessageRecord
-} from '@/tape/application/factPersistence'
-
-/** @deprecated Import Tape fact types from `@/tape/application/factPersistence`. */
-export type { TapeFactSource } from '@/tape/application/factPersistence'
