+++ test/main/memory/deepchatMemoryIngestionProjection.test.ts
@@ -1,12 +1,12 @@
 import { describe, expect, vi } from 'vitest'
-import { buildEffectiveTapeView } from '@/session/data/tapeEffectiveView'
+import { buildEffectiveTapeView } from '@/tape/domain/effectiveView'
 import { SqliteTapeLifecycleAdapter } from '@/tape/infrastructure/sqlite/tapeLifecycleAdapter'
 import { TOOL_SURFACE_TAPE_EVENT_NAMES } from '@/tape/domain/toolSurfaceFacts'
 import { TapeSkillMaterializationService } from '@/tape/application/skillMaterializationService'
 import { hashSkillEffectiveContent } from '@/tape/domain/skillMaterialization'
 import { Database, nativeSqliteItIf } from '../nativeSqliteHarness'
 
-const entriesModule = Database ? await import('@/session/data/tables/deepchatTapeEntries') : null
+const entriesModule = Database ? await import('@/tape/infrastructure/sqlite/tapeEntryStore') : null
 const projectionModule = Database
   ? await import('@/memory/data/tables/deepchatMemoryIngestionProjection')
   : null
